documents: refactoring action buttons
