+++ projects/admin/src/app/record/detail-view/document-detail-view/document-detail-view.component.ts
@@ -15,21 +15,15 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
-import { Record, RecordService } from '@rero/ng-core';
-import { DetailRecord } from '@rero/ng-core/lib/record/detail/view/detail-record';
+import { DetailRecord, RecordService } from '@rero/ng-core';
 import { IPermissions, PERMISSIONS, PermissionsService } from '@rero/shared';
-import { cloneDeep } from 'lodash-es';
-import { BsModalService } from 'ngx-bootstrap/modal';
 import { Observable, of, Subscription } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 import { DocumentApiService } from '../../../api/document-api.service';
-import { IdentifierTypes } from '../../../classes/identifiers';
 import { OperationLogsService } from '../../../service/operation-logs.service';
-import { DialogImportComponent } from './dialog-import/dialog-import.component';
 
 @Component({
   selector: 'admin-document-detail-view',
@@ -65,38 +59,12 @@ export class DocumentDetailViewComponent implements DetailRecord, OnInit, OnDest
   /** Enables or disables links */
   activateLink: boolean = true;
 
-  /** Mapping types for import */
-  private _mappingtypes = {
-    'bf:Ean': 'bf:Isbn',
-    'bf:Isbn': 'bf:Isbn',
-    'bf:Issn': 'bf:IssnL',
-    'bf:IssnL': 'bf:Issn'
-  };
-
-  /** Identifiers used to check for duplicates */
-  private _identifiersList = [
-    IdentifierTypes.ISBN,
-    IdentifierTypes.ISSN,
-    IdentifierTypes.DOI,
-    IdentifierTypes.LCCN,
-    IdentifierTypes.L_ISSN,
-    IdentifierTypes.EAN
-  ];
-
   /**
    * Is operation log enabled
    * @return boolean
    */
   get isEnabledOperationLog(): boolean {
     return this._operationLogsService.isLogVisible('documents');
-  }
-
-  /** Source for imported record. */
-  get source() {
-    if (this._activatedRouter.snapshot && this._activatedRouter.snapshot.params && this._activatedRouter.snapshot.params.type !== null) {
-      return this._activatedRouter.snapshot.params.type.replace('import_', '');
-    }
-    return null;
   }
 
   /** External identifier for imported record. */
@@ -122,7 +90,7 @@ export class DocumentDetailViewComponent implements DetailRecord, OnInit, OnDest
    * Show or hide local fields tab
    * @return boolean - if False, hide the local fields tab
    */
-  get showhideLocalFieldsTab(): boolean {
+  get showHideLocalFieldsTab(): boolean {
     return this._permissionsService.canAccess([PERMISSIONS.LOFI_SEARCH, PERMISSIONS.LOFI_CREATE]);
   }
 
@@ -132,8 +100,6 @@ export class DocumentDetailViewComponent implements DetailRecord, OnInit, OnDest
    * @param _activatedRouter - ActivatedRoute to get url parameters.
    * @param _recordService - RecordService to the MARC version for the record.
    * @param _operationLogsService - OperationLogsService
-   * @param _router - Router
-   * @param _bsModalService - BsModalService
    * @param _documentApiService - DocumentApiService
    * @param _permissionsService - PermissionsService
    */
@@ -142,8 +108,6 @@ export class DocumentDetailViewComponent implements DetailRecord, OnInit, OnDest
     private _activatedRouter: ActivatedRoute,
     private _recordService: RecordService,
     private _operationLogsService: OperationLogsService,
-    private _router: Router,
-    private _bsModalService: BsModalService,
     private _documentApiService: DocumentApiService,
     private _permissionsService: PermissionsService
   ) { }
@@ -187,73 +151,6 @@ export class DocumentDetailViewComponent implements DetailRecord, OnInit, OnDest
   }
 
   /**
-   * Import Document
-   * @param event - Event
-   * @param record - Current record
-   * @param data - Import source reference
-   */
-  importDocument(event: Event, record: any, data: { source: string, pid: string }): void {
-    event.preventDefault();
-    const rec = record.metadata;
-    const route = ['/records', 'documents', 'new'];
-    const queryParams = [];
-    // If we have identifiers
-    if (rec.identifiedBy) {
-      // We select only a defined part of the identifier
-      rec.identifiedBy
-        .filter((identifier: any) => this._identifiersList.includes(identifier.type))
-        .map((identifier: any) => {
-          queryParams.push(this._extractAndFormatQueryParams(identifier));
-          if (identifier.type in this._mappingtypes) {
-            const cidentifier = cloneDeep(identifier);
-            cidentifier.type = this._mappingtypes[cidentifier.type];
-            queryParams.push(this._extractAndFormatQueryParams(cidentifier));
-          }
-        });
-    }
-
-    // If we do not have a identifier and we have a title
-    // TODO: Nice to have -  create an backend API entrypoint to detect the duplicate resource.
-    // https://github.com/rero/rero-ils/issues/2900 (Generalize to all resources)
-    if (queryParams.length === 0 && rec.title) {
-      // We extract the title
-      const titles = rec.title.filter((rtitle: any) => rtitle.type === 'bf:Title' && '_text' in rtitle);
-      if (titles.length > 0) {
-        // We clean the text string by deleting some characters
-        const regex = /["\[\]]/gi;
-        queryParams.push(`title._text:"${titles[0]._text.replace(regex, '')}"`);
-      }
-    }
-
-    // No identifier and no title, we redirect to the add form.
-    if (queryParams.length === 0) {
-      this._router.navigate(route, { queryParams: data });
-    } else {
-      // Find documents(s) with query params
-      const query = queryParams.join(' OR ');
-      this._recordService.getRecords(
-        'documents', query, 1, undefined, undefined, undefined, { accept: 'application/rero+json' }
-      ).subscribe((response: Record) => {
-        if (this._recordService.totalHits(response.hits.total) === 0) {
-          this._router.navigate(route, { queryParams: data });
-        } else {
-          const config = {
-            initialState: {
-              records: response.hits.hits
-            }
-          };
-          const bsModalRef = this._bsModalService.show(DialogImportComponent, config);
-          bsModalRef.content.confirmation$.subscribe((confirmation: boolean) => {
-            if (confirmation) {
-              this._router.navigate(route, { queryParams: data });
-            }
-          });
-        }
-      });
-    }
-  }
-
-  /**
    * Format "part of" numbering for display
    *
    * @param num: numbering to format
@@ -293,21 +190,6 @@ export class DocumentDetailViewComponent implements DetailRecord, OnInit, OnDest
       default:
         return this._translateService.instant('Published in');
     }
-  }
-
-  /**
-   * Extract and format query params
-   * @param identifier - IdentifiedBy object
-   * @return string, query formatted
-   */
-  private _extractAndFormatQueryParams(identifier: any): string {
-    const query = [];
-    query.push(`identifiedBy.type:"${identifier.type}"`);
-    query.push(`identifiedBy.value:"${identifier.value}"`);
-    if (identifier.source) {
-      query.push(`identifiedBy.source:"${identifier.source}"`);
-    }
-    return `(${query.join(' AND ')})`;
   }
 
   /**
